requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
control==0.9.4
cycler==0.11.0
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.12.2
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
glfw==2.6.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
gymnasium==0.28.1
hydra-core==1.3.2
idna==3.4
jax-jumpy==1.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
mujoco==2.3.7
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pathtools==0.1.2
Pillow==10.0.0
platformdirs==3.10.0
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.3
PyOpenGL==3.1.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
sk8o_sim==0.1.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
sbx-rl==0.7.0
onnxruntime==1.15.1
onnx==1.14.1
smmap==5.0.0
stable-baselines3==2.0.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
torch==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.9
Werkzeug==2.3.6
